# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * منصة كلية تقنية المعلومات - IT College Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  message: string;
}

export interface SuccessMessage {
  message: string;
  success: boolean;
}

export interface Course {
  id: number;
  name: string;
  code: string;
  description?: string;
  year: number;
  semester: string;
  credits: number;
  instructor?: string;
  color?: string;
  createdAt?: string;
}

export interface CreateCourse {
  name: string;
  code: string;
  description?: string;
  year: number;
  semester: string;
  credits: number;
  instructor?: string;
  color?: string;
}

export interface Lecture {
  id: number;
  courseId: number;
  title: string;
  description?: string;
  lectureNumber: number;
  duration?: number;
  createdAt?: string;
  course?: Course;
}

export interface CreateLecture {
  courseId: number;
  title: string;
  description?: string;
  lectureNumber: number;
  duration?: number;
}

export type FileRecordType =
  (typeof FileRecordType)[keyof typeof FileRecordType];

export const FileRecordType = {
  pdf: "pdf",
  video: "video",
} as const;

export interface FileRecord {
  id: number;
  lectureId?: number;
  courseId?: number;
  title: string;
  filename: string;
  url: string;
  type: FileRecordType;
  size?: number;
  createdAt?: string;
  lecture?: Lecture;
  course?: Course;
}

export type ScheduleEntryType =
  (typeof ScheduleEntryType)[keyof typeof ScheduleEntryType];

export const ScheduleEntryType = {
  lecture: "lecture",
  lab: "lab",
  tutorial: "tutorial",
} as const;

export interface ScheduleEntry {
  id: number;
  courseId: number;
  /** 0=Sunday, 1=Monday, ..., 6=Saturday */
  dayOfWeek: number;
  startTime: string;
  endTime: string;
  location?: string;
  year: number;
  semester: string;
  type: ScheduleEntryType;
  course?: Course;
}

export type CreateScheduleEntryType =
  (typeof CreateScheduleEntryType)[keyof typeof CreateScheduleEntryType];

export const CreateScheduleEntryType = {
  lecture: "lecture",
  lab: "lab",
  tutorial: "tutorial",
} as const;

export interface CreateScheduleEntry {
  courseId: number;
  dayOfWeek: number;
  startTime: string;
  endTime: string;
  location?: string;
  year: number;
  semester: string;
  type: CreateScheduleEntryType;
}

export type GetCoursesParams = {
  year?: number;
  semester?: string;
};

export type GetLecturesParams = {
  courseId?: number;
};

export type GetFilesParams = {
  lectureId?: number;
  courseId?: number;
  type?: GetFilesType;
};

export type GetFilesType = (typeof GetFilesType)[keyof typeof GetFilesType];

export const GetFilesType = {
  pdf: "pdf",
  video: "video",
} as const;

export type UploadFileBodyType =
  (typeof UploadFileBodyType)[keyof typeof UploadFileBodyType];

export const UploadFileBodyType = {
  pdf: "pdf",
  video: "video",
} as const;

export type UploadFileBody = {
  file: Blob;
  lectureId?: number;
  courseId?: number;
  title: string;
  type: UploadFileBodyType;
};

export type GetScheduleParams = {
  year?: number;
  semester?: string;
};