# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: منصة كلية تقنية المعلومات - IT College Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: courses
    description: Course management
  - name: lectures
    description: Lecture management
  - name: files
    description: File management
  - name: schedule
    description: Schedule management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /courses:
    get:
      operationId: getCourses
      tags: [courses]
      summary: Get all courses
      parameters:
        - name: year
          in: query
          schema:
            type: number
        - name: semester
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a new course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourse"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"

  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a course by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Course found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: updateCourse
      tags: [courses]
      summary: Update a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourse"
      responses:
        "200":
          description: Course updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

  /lectures:
    get:
      operationId: getLectures
      tags: [lectures]
      summary: Get all lectures
      parameters:
        - name: courseId
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of lectures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lecture"
    post:
      operationId: createLecture
      tags: [lectures]
      summary: Create a new lecture
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLecture"
      responses:
        "201":
          description: Lecture created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"

  /lectures/{id}:
    get:
      operationId: getLecture
      tags: [lectures]
      summary: Get a lecture by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Lecture found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: updateLecture
      tags: [lectures]
      summary: Update a lecture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLecture"
      responses:
        "200":
          description: Lecture updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"
    delete:
      operationId: deleteLecture
      tags: [lectures]
      summary: Delete a lecture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

  /files:
    get:
      operationId: getFiles
      tags: [files]
      summary: Get all files
      parameters:
        - name: lectureId
          in: query
          schema:
            type: number
        - name: courseId
          in: query
          schema:
            type: number
        - name: type
          in: query
          schema:
            type: string
            enum: [pdf, video]
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileRecord"

  /files/upload:
    post:
      operationId: uploadFile
      tags: [files]
      summary: Upload a file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                lectureId:
                  type: number
                courseId:
                  type: number
                title:
                  type: string
                type:
                  type: string
                  enum: [pdf, video]
              required:
                - file
                - title
                - type
      responses:
        "201":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileRecord"

  /files/{id}:
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

  /schedule:
    get:
      operationId: getSchedule
      tags: [schedule]
      summary: Get schedule entries
      parameters:
        - name: year
          in: query
          schema:
            type: number
        - name: semester
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Schedule list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleEntry"
    post:
      operationId: createScheduleEntry
      tags: [schedule]
      summary: Create schedule entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleEntry"
      responses:
        "201":
          description: Schedule entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleEntry"

  /schedule/{id}:
    put:
      operationId: updateScheduleEntry
      tags: [schedule]
      summary: Update schedule entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleEntry"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleEntry"
    delete:
      operationId: deleteScheduleEntry
      tags: [schedule]
      summary: Delete schedule entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Error:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    SuccessMessage:
      type: object
      properties:
        message:
          type: string
        success:
          type: boolean
      required:
        - message
        - success

    Course:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        code:
          type: string
        description:
          type: string
        year:
          type: number
        semester:
          type: string
        credits:
          type: number
        instructor:
          type: string
        color:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - code
        - year
        - semester
        - credits

    CreateCourse:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
        year:
          type: number
        semester:
          type: string
        credits:
          type: number
        instructor:
          type: string
        color:
          type: string
      required:
        - name
        - code
        - year
        - semester
        - credits

    Lecture:
      type: object
      properties:
        id:
          type: number
        courseId:
          type: number
        title:
          type: string
        description:
          type: string
        lectureNumber:
          type: number
        duration:
          type: number
        createdAt:
          type: string
        course:
          $ref: "#/components/schemas/Course"
      required:
        - id
        - courseId
        - title
        - lectureNumber

    CreateLecture:
      type: object
      properties:
        courseId:
          type: number
        title:
          type: string
        description:
          type: string
        lectureNumber:
          type: number
        duration:
          type: number
      required:
        - courseId
        - title
        - lectureNumber

    FileRecord:
      type: object
      properties:
        id:
          type: number
        lectureId:
          type: number
        courseId:
          type: number
        title:
          type: string
        filename:
          type: string
        url:
          type: string
        type:
          type: string
          enum: [pdf, video]
        size:
          type: number
        createdAt:
          type: string
        lecture:
          $ref: "#/components/schemas/Lecture"
        course:
          $ref: "#/components/schemas/Course"
      required:
        - id
        - title
        - filename
        - url
        - type

    ScheduleEntry:
      type: object
      properties:
        id:
          type: number
        courseId:
          type: number
        dayOfWeek:
          type: number
          description: 0=Sunday, 1=Monday, ..., 6=Saturday
        startTime:
          type: string
        endTime:
          type: string
        location:
          type: string
        year:
          type: number
        semester:
          type: string
        type:
          type: string
          enum: [lecture, lab, tutorial]
        course:
          $ref: "#/components/schemas/Course"
      required:
        - id
        - courseId
        - dayOfWeek
        - startTime
        - endTime
        - year
        - semester
        - type

    CreateScheduleEntry:
      type: object
      properties:
        courseId:
          type: number
        dayOfWeek:
          type: number
        startTime:
          type: string
        endTime:
          type: string
        location:
          type: string
        year:
          type: number
        semester:
          type: string
        type:
          type: string
          enum: [lecture, lab, tutorial]
      required:
        - courseId
        - dayOfWeek
        - startTime
        - endTime
        - year
        - semester
        - type