# lib/api-zod/src/generated/types/createCourse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * منصة كلية تقنية المعلومات - IT College Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCourse {
  name: string;
  code: string;
  description?: string;
  year: number;
  semester: string;
  credits: number;
  instructor?: string;
  color?: string;
}