# lib/api-zod/src/generated/types/createScheduleEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * منصة كلية تقنية المعلومات - IT College Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateScheduleEntryType } from "./createScheduleEntryType";

export interface CreateScheduleEntry {
  courseId: number;
  dayOfWeek: number;
  startTime: string;
  endTime: string;
  location?: string;
  year: number;
  semester: string;
  type: CreateScheduleEntryType;
}