# lib/api-zod/src/generated/types/fileRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * منصة كلية تقنية المعلومات - IT College Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Course } from "./course";
import type { FileRecordType } from "./fileRecordType";
import type { Lecture } from "./lecture";

export interface FileRecord {
  id: number;
  lectureId?: number;
  courseId?: number;
  title: string;
  filename: string;
  url: string;
  type: FileRecordType;
  size?: number;
  createdAt?: string;
  lecture?: Lecture;
  course?: Course;
}